WIP: changing endpoints & api files location
