+++ client/src/containers/NodesList/NodesList.jsx
@@ -3,8 +3,8 @@ import { Link } from 'react-router-dom';
 import Header from '../Header';
 import Table from '../../components/Table';
 import * as constants from '../../utils/constants';
-import { get } from '../../services/api';
-import { uriNodes } from '../../services/endpoints';
+import { get } from '../../utils/api';
+import { uriNodes } from '../../utils/endpoints';
 
 class NodesList extends Component {
   state = {
